Add BaseMatFormField class JC-567
